Refactor some method sin EntityInsertPanel moving them into model classes where they make sense
